+++ src/Events/RunnerStarted.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace MyTester\Events;
+
+/**
+ * Triggers when automated tests runner started running
+ *
+ * @author Jakub Konečný
+ */
+final readonly class RunnerStarted
+{
+}